+++ src/hooks/useFunPuntoFijo.jsx
@@ -1,13 +1,13 @@
 import React, { useState } from 'react'
 
 export const useFunPuntoFijo = () => {
-    
+
     const [calculos, setCalculos] = useState([]);
     const [iscalculosfull, setIscalculosfull] = useState(false);
 
     const [number, setNumber] = useState(
         {
-            numeroIteraciones: '', 
+            numeroIteraciones: '',
             criterioTolerancia: '',
             IntervaloA: '',
             IntervaloB: ''
@@ -16,7 +16,7 @@ export const useFunPuntoFijo = () => {
 
     const handleInput = (event) => {
         const { name, value } = event.target
-        setNumber({...number, [name]: value})
+        setNumber({ ...number, [name]: value })
     }
 
     const Fun = x => {
@@ -30,47 +30,47 @@ export const useFunPuntoFijo = () => {
 
         const { numeroIteraciones, criterioTolerancia } = number
 
-        if(Number(numeroIteraciones) <= 0){
+        if (Number(numeroIteraciones) <= 0) {
             console.log('el criterio de tolerancia no puede ser cero')
             return
         }
-        else{
+        else {
             let n;
             let ea;
             let es;
             let xi;
             let Ar;
             let gx;
-                
+
             xi = 0;
-            es = Number(criterioTolerancia); 
+            es = Number(criterioTolerancia);
             Ar = 0;
             n = 1;
             gx = 0;
-           
+
             do {
-                gx = Fun(xi) //0 - 0
-                ea = Math.abs((gx - Ar) / gx) * 100 //null - (0 - 0) / 0 * 100
-                Ar = gx // 0 = 0
-    
+                gx = Fun(xi);
+                ea = Math.abs((gx - Ar) / gx) * 100;
+                Ar = gx;
+
                 const calc = {
                     n,
-                    xi, 
+                    xi,
                     gx,
                     ea
                 }
-                
-                setCalculos(calculos =>[...calculos, calc])
-                
-                xi = gx //1
-                n = n + 1
-    
-            } while (ea > es);  
+
+                setCalculos(calculos => [...calculos, calc])
+
+                xi = gx;
+                n = n + 1;
+
+            } while (ea > es);
             setIscalculosfull(true);
         }
     }
 
-    return({
+    return ({
         calculos,
         iscalculosfull,
         number,
